added keyword arguments to blitz() so that it can handle distutils arguments just like inline()
converted all exceptions to call throw_error().  throw_error(exception,msg) will set the Python exception values and then through an integer exception.  This is the only one that seems to work on Mandrake.  Gordon McMillan says it happens on other platforms also when the python has been linked with a C compiler instead of a C++ compiler (which is almost always the case).


git-svn-id: http://svn.scipy.org/svn/numpy/trunk@110 94b884b6-d6fd-0310-90d3-974f1d3f35e1