cleaned up a few examples.
added support for wxPython on Unix.  Configuration information is gathered
by running wx-config with the proper flags specified (--cxxflags, --ldflags,
--libs, --gl-libs) to find libraries and include directories needed.  I
wish we could query the wxPython module directly and ask it for this
information, as it would be much more robust.  Currently we look in the
/usr/lib/wxPython directory for wx stuff.  If it isn't there, we're out of
luck.  The wxPython development files are required for this stuff to work.

Windows is actually worse.  I've updated the config files to support
wx2.3.3.1.  Is there a way to discover build type (ansi/unicode) from the
wxPython library?  Right now, it is hard coded to use the Unicode version
of the DLLs.  As for macros and libraries, I borrowed them from the make
file in the wx2.3.3.1/src/make.env file.  The wxPython devel package must
be installed the c:\ directory for all of this to work.

If people use alternative directories in either case, they'll have to hand
edit the wx_spec file to deal with this.


git-svn-id: http://svn.scipy.org/svn/numpy/trunk@382 94b884b6-d6fd-0310-90d3-974f1d3f35e1