added flags to Sun compiler to make it build modules in the build/temp.os-xxx directory.  This involved adding -moddir/temp_dir and -M temp_dir
absoft 6.2 doesn't handle building mod files in the build directory using a flag
(or at least I didn't find it).  This is brain dead.  anyway, I added the
machinery to move the files to the build directory explicitly after each f90
compile.  It is smart enough to not move mod files that previously existed
 there.  I am a bit worried whether the absoft compiler accepts multiple -p
flags to specify where to look for module files.  This isn't a problem unless
a user also specifies a directory using the module_dirs argument.

The Sun compiler also needed the mvec library on our machine, so I've added it.
Hopefully it is OK to have on all systems.??


git-svn-id: http://svn.scipy.org/svn/numpy/trunk@432 94b884b6-d6fd-0310-90d3-974f1d3f35e1