added setup_extension() method to ext_module to return distutils.core.Extension object for a weave generated extension.  This is helpful in using weave extensions within setup.py files.
added logic inside ext_module.generate_module to check whether the
new extension module code is different from the code that may already 
exist in a previously generated extension module file.  If the code is
the same, the file is not overwritten.  This prevents setup.py from thinking
the module is always out of date and must be recompiled.


git-svn-id: http://svn.scipy.org/svn/numpy/trunk/scipy_distutils@483 94b884b6-d6fd-0310-90d3-974f1d3f35e1