fixed a typecode Float->float issue in average.
This is not pretty code it seems overly complicated, and it has an "eval"
statement in it where it builds a string and evaluates it.  I can't imagine
this is the best approach to this -- but I also am not going to spend the
time to examine it right now.

git-svn-id: http://svn.scipy.org/svn/numpy/trunk/numpy/lib@3699 94b884b6-d6fd-0310-90d3-974f1d3f35e1