Update new correlation support.
Rename PyArray_Acorrelate to PyArray_Correlate2 to follow the current NumPy
convention when adding new function and keeping the old one for backward
compatibility.

Add keyword to correlate to switch old/new behavior instead of adding
acorrelate, which is not discoverable.

Raise a DeprecationWarning when the old behavior is used - old behavior still
the default, though.

git-svn-id: http://svn.scipy.org/svn/numpy/trunk@7135 94b884b6-d6fd-0310-90d3-974f1d3f35e1