Change how we decorate npy_math functions.
For platforms without the necessary math functions, we used to first
implement them by ourselves, and then decorate the math function with
npy_ prefix, whether the undecorated functions are from the
implementation or from ourselves.

This is not always practical, because we may want to use our own
implementation even if the platform has one implementation. Instead,
when we define our own implementation, we directly define the decorated
function, and only decorate functions when we use the platform
implementation.

Example: say the function hypot is available on the platform, but buggy.
If we undefine HAVE_HYPOT, in the former scheme, we would have first:

  static double hypot(double x, double y);

And then:

  double npy_hypot(double x, double y);

But this would clash with the platform hypot. By defining directly
hpy_hypot, and put our implementation there, we avoid this clash.

git-svn-id: http://svn.scipy.org/svn/numpy/trunk@7145 94b884b6-d6fd-0310-90d3-974f1d3f35e1