Forward declare PyArrayIterObject
To fix the neighborhood iterator, we need to be able to add a member to
PyArrayIterObject which is a function pointer (which will do the
padding-specific translation coordinates -> data pointer: poor man's
polymorphism). As the function pointer takes an iter object as its first
argument, we need to make a proper declaration of PyArrayIterObject,
instead of using an anonymous struct as previously defined. I am not
sure whether this may break the ABI or not, though.

git-svn-id: http://svn.scipy.org/svn/numpy/trunk@7287 94b884b6-d6fd-0310-90d3-974f1d3f35e1