BUG: (#1221) special case mac os x in numpyconfig.h
Universal builds break the configure stage: we have to harcode the
arch-specific values in the case of mac os x, as we have only one configuration
stage per compile, but several arch per compile with the braindead -arch
machinery.

We rename the old numpyconfig.h to a private header, and numpyconfig.h itself
post-fix the values in the case of mac os x.

git-svn-id: http://svn.scipy.org/svn/numpy/trunk@7790 94b884b6-d6fd-0310-90d3-974f1d3f35e1