BUG: add a work-around for #1312 -- don't define bf_releasebuffer at all
The problem is that PyArg_ParseTuple("s#", ...) does not accept objects
with the new buffer interface, if they have bf_releasebuffer defined.

We now work around this by not defining bf_releasebuffer, and releasing
any info associated with the buffer interface when array or its
descriptor is deallocated.

The shape+strides information for arrays stays static due to the
refcounting lock. The format information for dtypes may change
if field names are changed on-the-fly.

XXX: dtype field name changes are not currently handled.

git-svn-id: http://svn.scipy.org/svn/numpy/trunk@7974 94b884b6-d6fd-0310-90d3-974f1d3f35e1