ENH: core: make PyArray_FromAny error message more clear
When building string/unicode arrays, PyArray_FromAny would try building
a 0-d array as a last resort. If the object is a finite-length sequence,
this always fails with an error "ValueError: setting an array element
with a sequence", as STRING/UNICODE_setitem do not allow this.

In some cases this message is confusing: for example on Py3,

    array([['a', 'b'], ['c', '\u03a3']], dtype='S2')

would fail because of this, even though the real error is that '\u03a3'
cannot be unambiguously cast to bytes().

git-svn-id: http://svn.scipy.org/svn/numpy/trunk@8144 94b884b6-d6fd-0310-90d3-974f1d3f35e1