ENH: implement PEP 3118 buffer consumer on Py2.6, adding a simple Memoryview object
To use PEP 3118 buffers as ndarray bases, we need to have a way to
release the buffers on garbage collection. The easiest way to do this is
to use a Memoryview object -- but they are not available on Python 2.6.

Hence, we implement a minimal memory view object that only handles the
garbage collection, and use it on Python 2.6.  It also implements the
new buffer protocol, to allow testing the functionality.

git-svn-id: http://svn.scipy.org/svn/numpy/trunk@8152 94b884b6-d6fd-0310-90d3-974f1d3f35e1